# app/src/main/java/com/vrerv/springboottemplate/server/feature/user/model/LoginModel.java
package com.vrerv.springboottemplate.server.feature.user.model;

import javax.validation.constraints.NotBlank;

import lombok.AccessLevel;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Data
@Builder
@AllArgsConstructor(access = AccessLevel.PACKAGE)
@NoArgsConstructor(access = AccessLevel.PACKAGE)
@Setter(AccessLevel.PRIVATE)
public class LoginModel {

	@NotBlank
	private String account;

	@NotBlank
	private String password;
}